Refactoring: own folder for use cases
