+++ demo/lib/main.dart
@@ -1,12 +1,11 @@
+import 'package:demo/data/theme_data.dart';
 import 'package:demo/data/use_case.dart';
-import 'package:demo/data/use_cases/contextual_controls/contextual_controls_use_case.dart';
-import 'package:demo/data/use_cases/filter/filter_use_case.dart';
-import 'package:demo/data/use_cases/navigation/navigation_use_case.dart';
 import 'package:demo/demo/demo_page.dart';
-import 'package:demo/theme_data.dart';
+import 'package:demo/use_cases/contextual_controls/contextual_controls_use_case.dart';
+import 'package:demo/use_cases/contextual_info/contextual_info_use_case.dart';
+import 'package:demo/use_cases/filter/filter_use_case.dart';
+import 'package:demo/use_cases/navigation/navigation_use_case.dart';
 import 'package:flutter/material.dart';
-
-import 'data/use_cases/contextual_info/contextual_info_use_case.dart';
 
 void main() => runApp(DemoApp());
