+++ demo/lib/use_cases/navigation/navigation_use_case.dart
@@ -1,5 +1,5 @@
 import 'package:demo/data/use_case.dart';
-import 'package:demo/data/use_cases/navigation/navigation.dart';
+import 'package:demo/use_cases/navigation/navigation.dart';
 
 /// Navigation use case description.
 class NavigationUseCase extends UseCase {
